Refactor EGraphLike and related classes to introduce ReadOnlyEGraph interface and improve structure
